requirements.txt
flask==3.0.0
flask-socketio==5.3.6
tensorflow==2.15.0
opencv-python-headless==4.8.1.78
numpy==1.26.2
matplotlib==3.8.2
scikit-learn==1.3.2
Pillow==10.1.0
pandas==2.1.4
seaborn==0.13.0
eventlet==0.33.3
werkzeug==3.0.1
h5py==3.10.0
tqdm==4.66.1
plotly==5.18.0
gunicorn==21.2.0
